SAMZA-2084: [Samza Sql] add logging when exceptions are swallowed and rename util package
