+++ samza-sql/src/main/java/org/apache/samza/sql/util/ReflectionUtils.java
@@ -17,7 +17,7 @@
 * under the License.
 */
 
-package org.apache.samza.sql.testutil;
+package org.apache.samza.sql.util;
 
 import java.lang.reflect.Constructor;
 import java.util.stream.IntStream;
