+++ samza-sql/src/main/java/org/apache/samza/sql/util/SamzaSqlQueryParser.java
@@ -17,7 +17,7 @@
 * under the License.
 */
 
-package org.apache.samza.sql.testutil;
+package org.apache.samza.sql.util;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
