+++ samza-sql/src/test/java/org/apache/samza/sql/util/TestMetricsRegistryImpl.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.samza.sql.testutil;
+package org.apache.samza.sql.util;
 
 import java.util.ArrayList;
 import java.util.HashMap;
